Extract strength as int32 in SignalPoll method_call's reply
WPA Supplicant sends "rssi" as int32 in SignalPoll's method_call reply.
So strength should be int32 instead of unsigned char variable.

========================
method call time=1570774928.393859 sender=:1.24 -> destination=fi.w1.wpa_supplicant1 serial=364 path=/fi/w1/wpa_supplicant1/Interfaces/0; interface=fi.w1.wpa_supplicant1.Interface; member=SignalPoll
method return time=1570774928.395629 sender=:1.47 -> destination=:1.24 serial=327 reply_serial=364
   variant       array [
         dict entry(
            string "rssi"
            variant                int32 -60
         )
         dict entry(
            string "linkspeed"
            variant                int32 24
         )
         dict entry(
            string "noise"
            variant                int32 9999
         )
         dict entry(
            string "frequency"
            variant                uint32 2452
         )
      ]
========================

+++ gsupplicant/gsupplicant.h
@@ -296,7 +296,7 @@ typedef void (*GSupplicantInterfaceCallback) (int result,
 
 #if defined TIZEN_EXT
 typedef void (*GSupplicantMaxSpeedCallback) (int result, int maxspeed,
-					     uint8_t strength, void *user_data);
+					     int strength, void *user_data);
 #endif
 
 void g_supplicant_interface_cancel(GSupplicantInterface *interface);
